Added footer module, general cleanup.
Added the Bootstrap module for a sticky footer at the bottom of each
page - you can use it on your own extra pages by copying the including
it via PHP from “includes” directory. :)

Also, cleaned up the introductory section a little more, with more
modifications to come. In the next update, links to our web locations
will be fixed. This means that if example text says “such and such…
LINK TO A SITE WE USE … such and such”, it will lead there - this could
be useful for help with writing into your own CorsoDB App :)