Minor fixes, New features in testing...
This update features some minor fixes around the framework, but the
main point of this release is the new feature experiments we’re
conducting.

We’ve added more user features to the framework - most of this is bare
design shell, with a need for code to function. We’re just
experimenting with these features, and we may choose to scrap them at
any time - what’s important is that we’re testing new ideas all the
time :)

Cheers guys,
Deano